Use include_bytes! instead of incbin.
This is the second attempt at this patch. The first version was reverted
in 3547710921003ebdb32bc83f7e5eeda426289214

The problem, I suspect, was that the snapshot was represented as a
source_set, which inserted a node into the dependency tree.
include_bytes does properly insert the snapshot into rustc's depfile but
the use of source_set confused gn. Now the that the deno executable has
the create_deno_snapshot as a direct dependency, changes will be
propagated.